Factory reset, fixes, quality of life
- Completed testing of factory reset and reworked a few bits of it
- Fixed various issues turned up during testing
- Implemented a variety of quality-of-life updates

Factory reset test blocked by a synchronization issue with MatterDevice with no obvious workaround so I probably need to
address.  Test is commented out for now, probably would pass otherwise.

+++ packages/matter.js/src/behavior/state/transaction/Tx.ts
@@ -69,23 +69,23 @@ export function executeTransaction<T>(
                         throw error;
                     }
 
-                    let promise = MaybePromise.then(
+                    return MaybePromise.then(
                         () => {
                             logger.warn("Rolling back", tx.via, "due to error:", error);
                             return tx.rollback();
                         },
-                        undefined,
+                        () => {
+                            tx.destroy();
+                            throw error;
+                        },
                         error2 => {
-                            logger.error("Secondary error in", tx.via, "rollback:", error2);
+                            if (error !== error2) {
+                                logger.error("Secondary error in", tx.via, "rollback:", error2);
+                                tx.destroy();
+                            }
+                            throw error;
                         },
                     );
-
-                    promise = MaybePromise.finally(promise, () => {
-                        tx.destroy();
-                        throw error;
-                    });
-
-                    return promise as MaybePromise<T>;
                 },
             ),
